refactor: Move server to cmd
because it's part of the cli and serves both the api endpoints and the
web interface
