+++ internal/cli/server.go
@@ -1,12 +1,18 @@
 package cli
 
 import (
+	"fmt"
 	"net"
+	"net/http"
+	"os"
+	"path"
 	"strconv"
 
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
-	"github.com/suse/carrier/internal/api"
+	apiv1 "github.com/suse/carrier/internal/api/v1"
+	"github.com/suse/carrier/internal/filesystem"
+	"github.com/suse/carrier/internal/web"
 )
 
 func init() {
@@ -26,8 +32,39 @@ var CmdServer = &cobra.Command{
 			return err
 		}
 
-		return api.StartServer(listener)
+		// TODO: Use `ui` package
+		fmt.Println("listening on", listener.Addr().String())
+
+		http.Handle("/api/v1/", logRequestHandler(apiv1.Router()))
+		http.Handle("/", logRequestHandler(web.Router()))
+		// Static files
+		var assetsDir http.FileSystem
+		if os.Getenv("LOCAL_FILESYSTEM") == "true" {
+			assetsDir = http.Dir(path.Join(".", "embedded-web-files", "assets"))
+		} else {
+			assetsDir = filesystem.Assets()
+		}
+		http.Handle("/assets/", http.StripPrefix("/assets/", http.FileServer(assetsDir)))
+
+		return http.Serve(listener, nil)
 	},
 	SilenceErrors: true,
 	SilenceUsage:  true,
+}
+
+// loggingmiddleware for requests
+func logRequestHandler(h http.Handler) http.Handler {
+	fn := func(w http.ResponseWriter, r *http.Request) {
+
+		// call the original http.Handler
+		h.ServeHTTP(w, r)
+
+		// log the request
+		uri := r.URL.String()
+		method := r.Method
+		// TODO: Use verbosity level to decide if we print or not
+		fmt.Printf("%s %s\n", method, uri)
+	}
+
+	return http.HandlerFunc(fn)
 }
